Creating a bluetooth interface
